+++ app/services/product_service.rb
@@ -50,7 +50,7 @@ class ProductService
         }
       end || {}
       product.stock = remaining_stock
-      notify_on_low_stock(product) if product.low_stock_alert
+      notify_on_low_stock(product, stock_movements)
       product
     end
 
@@ -187,11 +187,13 @@ class ProductService
       movements.where(reason: filters[:reason])
     end
 
-    def notify_on_low_stock(product)
+    def notify_on_low_stock(product, stock_movements = nil)
+      return product unless product.low_stock_alert
       return product unless product.low_stock_threshold
 
-      if (product.stock['internal'] <= product.low_stock_threshold) ||
-         (product.stock['external'] <= product.low_stock_threshold)
+      affected_stocks = stock_movements&.map { |m| m[:stock_type] }&.uniq
+      if (product.stock['internal'] <= product.low_stock_threshold && affected_stocks&.include?('internal')) ||
+         (product.stock['external'] <= product.low_stock_threshold && affected_stocks&.include?('external'))
         NotificationCenter.call type: 'notify_admin_low_stock_threshold',
                                 receiver: User.admins_and_managers,
                                 attached_object: product